Added some logic with creating new orders, added products.

+++ src/main/java/com/paulok777/entity/OrderProducts.java
@@ -24,5 +24,7 @@ public class OrderProducts {
     @JoinColumn(name = "product_id")
     Product product;
 
-    int amount;
+    long amount;
+
+
 }
